fix e2e test

+++ tests/multicard/test_qwen3_moe.py
@@ -28,7 +28,7 @@ def test_models_distributed_Qwen3_MOE_Aclgraph_TP2():
     example_prompts = [
         "Hello, my name is",
     ]
-    dtype = "half"
+    dtype = "bfloat16"
     max_tokens = 5
     with VllmRunner(
             "Qwen/Qwen3-30B-A3B",
